Small fix of previous commit
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@5273 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/Schema.cpp
@@ -677,7 +677,7 @@ BOOL CSchema::Validate(CXMLElement* pXML, BOOL bFix)
 		}
 		else if ( pMember->m_bYear )
 		{
-			short nYear = 0;
+			int nYear = 0;
 			if ( _stscanf( str, L"%i", &nYear ) != 1 ) return FALSE;
 			if ( nYear < 1000 || nYear > 9999 ) return FALSE;
 		}
